refactor(editor): Extract usage code into features
