+++ packages/frontend/editor-ui/src/features/evaluation.ee/components/Paywall/EvaluationsPaywall.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { useI18n } from '@n8n/i18n';
-import { COMMUNITY_PLUS_ENROLLMENT_MODAL } from '@/features/communityPlus/communityPlus.constants';
+import { COMMUNITY_PLUS_ENROLLMENT_MODAL } from '@/features/usage/usage.constants';
 import { useUIStore } from '@/stores/ui.store';
 
 import { N8nActionBox } from '@n8n/design-system';
