+++ packages/frontend/editor-ui/src/features/usage/components/CommunityPlusEnrollmentModal.test.ts
@@ -3,8 +3,8 @@ import userEvent from '@testing-library/user-event';
 import { mockedStore } from '@/__tests__/utils';
 import { createComponentRenderer } from '@/__tests__/render';
 import CommunityPlusEnrollmentModal from './CommunityPlusEnrollmentModal.vue';
-import { COMMUNITY_PLUS_ENROLLMENT_MODAL } from '../communityPlus.constants';
-import { useUsageStore } from '@/stores/usage.store';
+import { COMMUNITY_PLUS_ENROLLMENT_MODAL } from '../usage.constants';
+import { useUsageStore } from '../usage.store';
 import { useToast } from '@/composables/useToast';
 import { useTelemetry } from '@/composables/useTelemetry';
 import { useUsersStore } from '@/features/users/users.store';
