+++ packages/frontend/editor-ui/src/features/usage/components/CommunityPlusEnrollmentModal.vue
@@ -3,11 +3,11 @@ import { ref } from 'vue';
 import { createEventBus } from '@n8n/utils/event-bus';
 import type { Validatable, IValidator } from '@n8n/design-system';
 import { VALID_EMAIL_REGEX } from '@/constants';
-import { COMMUNITY_PLUS_DOCS_URL } from '../communityPlus.constants';
+import { COMMUNITY_PLUS_DOCS_URL } from '../usage.constants';
 import Modal from '@/components/Modal.vue';
 import { useI18n } from '@n8n/i18n';
 import { useToast } from '@/composables/useToast';
-import { useUsageStore } from '@/stores/usage.store';
+import { useUsageStore } from '../usage.store';
 import { useTelemetry } from '@/composables/useTelemetry';
 import { useUsersStore } from '@/features/users/users.store';
