+++ packages/frontend/editor-ui/src/features/usage/usage.test.ts
@@ -1,5 +1,5 @@
 import { createPinia, setActivePinia } from 'pinia';
-import { useUsageStore } from '@/stores/usage.store';
+import { useUsageStore } from '@/features/usage/usage.store';
 
 describe('Usage and plan store', () => {
 	beforeEach(() => {
